+++ bin/webpack-dev-server.js
@@ -15,13 +15,18 @@ const setupExitSignals = require('../lib/utils/setupExitSignals');
 const colors = require('../lib/utils/colors');
 const processOptions = require('../lib/utils/processOptions');
 const createLogger = require('../lib/utils/createLogger');
-const findPort = require('../lib/utils/findPort');
 const getVersions = require('../lib/utils/getVersions');
 const options = require('./options');
 
 let server;
-
-setupExitSignals(server);
+const serverData = {
+  server: null,
+};
+// we must pass an object that contains the server object as a property so that
+// we can update this server property later, and setupExitSignals will be able to
+// recognize that the server has been instantiated, because we will set
+// serverData.server to the new server object.
+setupExitSignals(serverData);
 
 // Prefer the local installation of webpack-dev-server
 if (importLocal(__filename)) {
@@ -99,6 +104,7 @@ function startDevServer(config, options) {
 
   try {
     server = new Server(compiler, options, log);
+    serverData.server = server;
   } catch (err) {
     if (err.name === 'ValidationError') {
       log.error(colors.error(options.stats.colors, err.message));
@@ -148,18 +154,11 @@ function startDevServer(config, options) {
       });
     });
   } else {
-    findPort(options.port)
-      .then((port) => {
-        options.port = port;
-        server.listen(options.port, options.host, (err) => {
-          if (err) {
-            throw err;
-          }
-        });
-      })
-      .catch((err) => {
+    server.listen(options.port, options.host, (err) => {
+      if (err) {
         throw err;
-      });
+      }
+    });
   }
 }