+++ test/e2e/Progress.test.js
@@ -5,20 +5,34 @@
 */
 const fs = require('fs');
 const { resolve } = require('path');
-const testServer = require('../helpers/test-server');
-const reloadConfig = require('../fixtures/reload-config/webpack.config');
+const reloadConfig = require('../fixtures/reload-config-2/webpack.config');
 const runBrowser = require('../helpers/run-browser');
 const port = require('../ports-map').Progress;
+const {
+  reloadReadyDelay,
+  completeReloadDelay,
+} = require('../helpers/puppeteer-constants');
 
-const cssFilePath = resolve(__dirname, '../fixtures/reload-config/main.css');
+const cssFilePath = resolve(__dirname, '../fixtures/reload-config-2/main.css');
 
 describe('client progress', () => {
+  let testServer;
+  let processStderrMock;
+
   describe('using hot', () => {
     beforeAll((done) => {
+      // ProgressPlugin uses process.stderr and reset webpack
+      jest.resetModules();
+      processStderrMock = jest
+        .spyOn(process.stderr, 'write')
+        .mockImplementation();
+      testServer = require('../helpers/test-server');
+
       fs.writeFileSync(
         cssFilePath,
         'body { background-color: rgb(0, 0, 255); }'
       );
+
       const options = {
         port,
         host: '0.0.0.0',
@@ -26,15 +40,17 @@ describe('client progress', () => {
         hot: true,
         progress: true,
         watchOptions: {
-          poll: 500,
+          poll: true,
         },
       };
+
       testServer.startAwaitingCompilation(reloadConfig, options, done);
     });
 
     afterAll((done) => {
       fs.unlinkSync(cssFilePath);
       testServer.close(done);
+      processStderrMock.mockRestore();
     });
 
     describe('on browser client', () => {
@@ -42,20 +58,22 @@ describe('client progress', () => {
         runBrowser().then(({ page, browser }) => {
           const res = [];
           page.waitForNavigation({ waitUntil: 'load' }).then(() => {
-            fs.writeFileSync(
-              cssFilePath,
-              'body { background-color: rgb(255, 0, 0); }'
-            );
-            page.waitFor(10000).then(() => {
-              browser.close().then(() => {
-                // check that there is some percentage progress output
-                const regExp = /^\[WDS\] [0-9]{1,3}% - /;
-                const match = res.find((line) => {
-                  return regExp.test(line);
+            page.waitFor(reloadReadyDelay).then(() => {
+              fs.writeFileSync(
+                cssFilePath,
+                'body { background-color: rgb(255, 0, 0); }'
+              );
+              page.waitFor(completeReloadDelay).then(() => {
+                browser.close().then(() => {
+                  // check that there is some percentage progress output
+                  const regExp = /^\[WDS\] [0-9]{1,3}% - /;
+                  const match = res.find((line) => {
+                    return regExp.test(line);
+                  });
+                  // eslint-disable-next-line no-undefined
+                  expect(match).not.toEqual(undefined);
+                  done();
                 });
-                // eslint-disable-next-line no-undefined
-                expect(match).not.toEqual(undefined);
-                done();
               });
             });
           });