+++ test/options.test.js
@@ -12,6 +12,16 @@ const config = require('./fixtures/simple-config/webpack.config');
 describe('options', () => {
   jest.setTimeout(20000);
 
+  let consoleMock;
+
+  beforeAll(() => {
+    consoleMock = jest.spyOn(console, 'warn').mockImplementation();
+  });
+
+  afterAll(() => {
+    consoleMock.mockRestore();
+  });
+
   it('should match properties and errorMessage', () => {
     const properties = Object.keys(options.properties);
     const messages = Object.keys(options.errorMessage.properties);
@@ -136,10 +146,6 @@ describe('options', () => {
           'warning',
         ],
         failure: ['whoops!'],
-      },
-      clientMode: {
-        success: ['sockjs', require.resolve('../client/clients/SockJSClient')],
-        failure: [false],
       },
       compress: {
         success: [true],
@@ -351,14 +357,6 @@ describe('options', () => {
         success: [true],
         failure: [''],
       },
-      serverMode: {
-        success: [
-          'sockjs',
-          require.resolve('../lib/servers/SockJSServer'),
-          SockJSServer,
-        ],
-        failure: [false],
-      },
       serverSideRender: {
         success: [true],
         failure: [''],
@@ -399,6 +397,62 @@ describe('options', () => {
           'verbose',
         ],
         failure: ['whoops!', null],
+      },
+      transportMode: {
+        success: [
+          'ws',
+          'sockjs',
+          {
+            transportMode: {
+              server: 'sockjs',
+            },
+          },
+          {
+            transportMode: {
+              server: require.resolve('../lib/servers/SockJSServer'),
+            },
+          },
+          {
+            transportMode: {
+              server: SockJSServer,
+            },
+          },
+          {
+            transportMode: {
+              client: 'sockjs',
+            },
+          },
+          {
+            transportMode: {
+              client: require.resolve('../client/clients/SockJSClient'),
+            },
+          },
+          {
+            transportMode: {
+              server: SockJSServer,
+              client: require.resolve('../client/clients/SockJSClient'),
+            },
+          },
+        ],
+        failure: [
+          'nonexistent-implementation',
+          null,
+          {
+            transportMode: {
+              notAnOption: true,
+            },
+          },
+          {
+            transportMode: {
+              server: false,
+            },
+          },
+          {
+            transportMode: {
+              client: () => {},
+            },
+          },
+        ],
       },
       useLocalIp: {
         success: [false],